Rendered data from repo to display movie details

+++ android_app/app/src/main/java/com/example/movienight20/ui/MoviesListScreen.kt
@@ -18,7 +18,7 @@ import com.example.movienight20.ui.theme.PopularMoviesViewModel
 
 @Composable
 fun MoviesListScreen(
-    onClickMovieListItem: () -> Unit,
+    onClickMovieListItem: (Int) -> Unit,
     viewModel: PopularMoviesViewModel
 ) {
     val viewStates by viewModel.viewState.collectAsState()
